Add hero section
Added a hero section in the html code. This is the first section
the user reads most of the time.

This section is going to be layout together with the header in an
flexbox. Differentiating the hero and header from the main content
and footer. As well as layed out seperately from the header in another
flexbox.

Daan Rosies (S5631742)